tuple.index, set.__eq__ and __ne__
Also fix str.rpartition, and add a signal handler for SIGUSR1 that
prints out a stack trace.  I wanted this because our stack unwinder
doesn't work when called from a gdb print command, so it was hard to
debug why we were in an infinite loop.

Also, fix the '-i' argument so that we get back to the repl if the
program throws an exception.