exec and eval compilations are always areGlobalsFromModule() == false
Before, if we saw that they were going to immediately be executed in
the current globals scope (ie no "in" clause), we would execute them
without custom globals.

In order to split the compilation from the execution, just always
compile them with custom globals support.  This involved some refactoring
of the assumptions about when the globals object gets passed or not.