Don't cache analysis results
On the simple test of `pyston -c "import pip"`, it reduces memory
usage from 194MB to 153MB (20% of the memory previously was cached
analysis data).  It also increases our benchmark geomean by ~1%,
which isn't great, but I think we can get that back eventually and
I don't think it's worth blocking this memory improvement for that.