increase number of generic ICs slots
also add some megamorphic microbenchmarks.  The code changes in this commit fix
iteration-megamorphic1.py. This fixes some of the easiest megamorphic IC's we're
seeing in the django template test.

iteration-megamorphic2.py is harder for us since we get a new class every time. If
we can figure out a way to guard on just the mro (and guarantee that the mro is shared
across all subclasses that don't specify their own), the we should still have 2 slots
used in iteration-megamorphic2.py