add "shapes" for BoxedClass
in order to make typeLookup not emit assembly for a given IC slot specialized to
a specific class's mro we need something that is shared between classes.

That something is a 'shape'.  Each class computes its own "local shape", which is
just the sha256 of its attribute names (in ascending order of slot offset).  This
local shape, combined with a parent's total shape, gives a subclass's total shape.

This is similar to the way hidden classes work, but with shapes we want a single
identifier (the sha256 hash) that describes multiple attributes, as opposed to a
separate hidden class for each attribute.

Once we've computed this total shape for a class, we can cache in the IC the index
of the class which has this attribute.  We also should know at this point what the
offset of the actual attribute within this class is.