Copy all HAVE_FOO defines from cpython
Previously we were only copying them when we noticed they were needed,
but we've been spending a fair bit of time hunting down issues that
are caused by a missing HAVE_ define.

So instead, bulk-copy the ones that the CPython configure script determined.
I bet we'll start running into some that were available on my 14.04
machine but won't be on other configurations; I'm hoping that an extra
define is easier to debug than a missing one (build error vs run error),
making this the better approach.