invalidate type shapes on delattr/setattr
the invalidation in setattr is done even if we set an existing attribute (as
opposed to adding a new one), something worth fixing.  The type's local shape is
invalidated at the point where the change happens, and we use PyType_Modified
to walk the tree of subclasses causing them to each recompute their total shape.

When hashing, write the attribute name sizes interleaved with the attribute names
to keep string concatenation from creating false equivalence.