Add alloc stats for more things
Add alloc stats to the base Box::operator new(), which is where non-builtin
types get allocated.

Also add stats for non-Python allocations (typically, conservative memory);
turns out that we are allocating a lot of conservative memory (both in terms
of allocations but also in terms of allocation size) for dicts.