libunwind: disable signal-blocking & conservative checks for performance
The signal-blocking is used to prevent libunwind from being entered
recursively, if a signal handler invokes libunwind. We don't call
libunwind from signal handlers, so this is not a problem for us.

The conservative checks just make crashes behave more nicely. They have
a small but measurable performance impact. We enable them in debug mode
and disable them for release mode.