Finalizer support for old-style classes.
Old-style classes always do an attribute lookup for __del__. It would
be slow to do attribute lookups in the GC all the time so for now, we
just assume all old-style classes have an ordered finalizer until the
finalizer is actually called.

Also, this commit refactors/cleans up the attribute lookup methods of
old style classes a little bit.