Temporarily disable type speculation
Unfortunately our type speculation strategy has bit-rot
over time, to the point that it's not super helpful.
Deopt (the flip side of the speculation) also has some bugs
in it which I was starting to trigger, so just disable the
whole thing for now.

Turning it off should just be a temporary measure and I'll
look soon into what we need to do to get it back on.