Helper for pinning down where time in a stattimer happens
If you define INVESTIGATE_STAT_TIMER to the name of the timer you want
to investigate, we will set an itimer that raises a SIGTRAP if you are
in that particular timer, but ignores the signal otherwise.  There's no
tooling on top of it, but just running that inside gdb is already helpful.