Revert to using gettimeofday to have measurements in us instead of ticks.
Also removing logic in Stats::dump that counts the total number of ticks
in timers. It relies on specific timer naming conventions that aren't
currently being followed. If, at a later time, we are interested in
having the sum of timer ticks, we should refactor the code to have a
separate class for counters that log time.