Retain the type object of instances when needed in the sweep phase.
This fixes a rare and subtle bug that occurs when both the last
instance of a class and the class itself (BoxedClass) get freed in
the sweep phase. If the BoxedClass is freed first, then the instance
looks at an invalid pointer when it tries to check if it has a
finalizer that should get called.