Optimization - remove the extra heap traversal of every collection.
We had an extra heap traversal to find objects with finalizers and
mark the classes of all objects in the heap.

Now, we instead maintain a list of objects with finalizers as they
get allocated.

To deal with the edge case of class objects getting freed before the
last instance of an object, we still have the logic of a heap traversal
that marks the classes of objects on the heap. However, it only gets
used during a garbage collection after a set number of classes get
allocated, which should be almost never.