Don't have cheetah_test look for system cheetah
cheetah's tests do a PATH search for the "cheetah" executable,
and since we were appending the test-cheetah's location to PATH,
it would find the system cheetah preferentially to the test cheetah.

This is fine since on most systems there isn't a system cheetah, but
somehow mine has one.  The system cheetah will invoke the system python,
which will end up passing more of the tests.

We could prepend PATH instead of append to it, but it looks like the
the tests don't need anything else, so for now don't pass in anything else
in PATH.