gc less frequently, and make collections faster
This change adds two knobs to configure gc frequency - heap occupancy percentage, and time
between collections.

The idea is that the collector tries to collect no more frequently than every QUICK_GC_FREQUENCY(500)
milliseconds.  If it NUM_QUICK_GC(2) collections in a row, it increases the size of the heap.

We also resize the heap if after a collection memory usage is above a certain threshold (right now it's
the same percentage as what triggered the collection.  should probably tune this more when/if we get
benchmarks when a lot of live objects after a collection.)

further work might be to shrink the heap if collections start taking too much time.