gc less frequently
simpler logic for determining gc frequency, based on how long it took to collect.

we optimistically increase the amount of headroom we give the mutator after each collection,
except if the collection took more than 50ms, at which point we decrease the headroom.

In practice this causes the collector to oscillate around that desired time.