Need to reload caller-save live-outs on the fast-paths
This is a very rare case, but I'm not sure how we didn't get burned
by this before.  But when llvm passes a caller-save register
as a live-out, we spill it around the patchpoint.  But I guess the
fastpath would just skip past the reloading.  This commit modifies
initializePatchpoint3, which handles this, to also pass back the
continue_addr in addition to some other metadata about the spilling/
reloading it had to do.