Tests for finalization code.
New testing helper function ensures that objects are garbage
collected and their finalizers tested in a more strict order. This
reduces the chances that an object isn't collected because a reference
remains in the stack, but it's hard to fully solve this problem as long
as we have conservative stack scanning.