Don't call getGlobalsModule() when creating a function
This causes us to do some unwinding which ends up showing up
in the stats.  We have the globals dict available, so just use
that.

Things are a bit complicated because we use "globals==NULL" to
signify certain things, but the asserts seem to be panning out so
I think it's ok.